refactor(chat-ui): wire flutterui to consume bond_chat_ui package
Replace inline chat rendering code with re-export stubs pointing to
bond_chat_ui. Add BondChatMessageItem Riverpod wrapper that bridges
provider-based services (threadService, fileService, agentDetails) to
the package's callback API. Remove dead clipboard helper platform files.
