+++ flutterui/lib/presentation/screens/chat/widgets/feedback_dialog.dart
@@ -1,229 +1,2 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-
-class FeedbackDialog extends StatefulWidget {
-  final String feedbackType;
-  final String? existingMessage;
-  final bool isEditing;
-  final VoidCallback onCancel;
-  final Function(String feedbackType, String? message) onSubmit;
-  final VoidCallback? onDelete;
-
-  const FeedbackDialog({
-    super.key,
-    required this.feedbackType,
-    this.existingMessage,
-    this.isEditing = false,
-    required this.onCancel,
-    required this.onSubmit,
-    this.onDelete,
-  });
-
-  @override
-  State<FeedbackDialog> createState() => _FeedbackDialogState();
-}
-
-class _FeedbackDialogState extends State<FeedbackDialog>
-    with SingleTickerProviderStateMixin {
-  late final TextEditingController _controller;
-  late final FocusNode _textFieldFocusNode;
-  late final FocusNode _keyboardListenerFocusNode;
-  late final AnimationController _animationController;
-  late final Animation<double> _slideAnimation;
-  late final Animation<double> _fadeAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = TextEditingController(text: widget.existingMessage ?? '');
-    _textFieldFocusNode = FocusNode();
-    _keyboardListenerFocusNode = FocusNode();
-
-    _animationController = AnimationController(
-      duration: const Duration(milliseconds: 200),
-      vsync: this,
-    );
-
-    _slideAnimation = Tween<double>(begin: -10.0, end: 0.0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.easeOut),
-    );
-
-    _fadeAnimation = Tween<double>(begin: 0.0, end: 1.0).animate(
-      CurvedAnimation(parent: _animationController, curve: Curves.easeOut),
-    );
-
-    _animationController.forward();
-
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _textFieldFocusNode.requestFocus();
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    _textFieldFocusNode.dispose();
-    _keyboardListenerFocusNode.dispose();
-    _animationController.dispose();
-    super.dispose();
-  }
-
-  void _handleSubmit() {
-    widget.onSubmit(widget.feedbackType, _controller.text.trim().isEmpty ? null : _controller.text.trim());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final isThumbUp = widget.feedbackType == 'up';
-
-    return AnimatedBuilder(
-      animation: _animationController,
-      builder: (context, child) {
-        return Transform.translate(
-          offset: Offset(0, _slideAnimation.value),
-          child: Opacity(
-            opacity: _fadeAnimation.value,
-            child: child,
-          ),
-        );
-      },
-      child: KeyboardListener(
-        focusNode: _keyboardListenerFocusNode,
-        onKeyEvent: (event) {
-          if (event is KeyDownEvent &&
-              event.logicalKey == LogicalKeyboardKey.escape) {
-            widget.onCancel();
-          }
-        },
-        child: Container(
-          margin: const EdgeInsets.only(top: 8),
-          padding: const EdgeInsets.all(12),
-          decoration: BoxDecoration(
-            color: colorScheme.surfaceContainerHighest,
-            borderRadius: BorderRadius.circular(12),
-            boxShadow: [
-              BoxShadow(
-                color: colorScheme.shadow.withAlpha(26),
-                blurRadius: 8,
-                offset: const Offset(0, 2),
-              ),
-            ],
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Row(
-                children: [
-                  Icon(
-                    isThumbUp ? Icons.thumb_up : Icons.thumb_down,
-                    size: 16,
-                    color: colorScheme.primary,
-                  ),
-                  const SizedBox(width: 8),
-                  Text(
-                    widget.isEditing ? 'Edit feedback' : 'Add feedback',
-                    style: TextStyle(
-                      fontSize: 13,
-                      fontWeight: FontWeight.w500,
-                      color: colorScheme.onSurface,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              TextField(
-                controller: _controller,
-                focusNode: _textFieldFocusNode,
-                maxLines: 2,
-                minLines: 1,
-                style: TextStyle(
-                  fontSize: 14,
-                  color: colorScheme.onSurface,
-                ),
-                decoration: InputDecoration(
-                  hintText: 'Why did you choose this? (optional)',
-                  hintStyle: TextStyle(
-                    fontSize: 14,
-                    color: colorScheme.onSurfaceVariant.withAlpha(153),
-                  ),
-                  border: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8),
-                    borderSide: BorderSide(color: colorScheme.outline),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8),
-                    borderSide: BorderSide(color: colorScheme.outline),
-                  ),
-                  focusedBorder: OutlineInputBorder(
-                    borderRadius: BorderRadius.circular(8),
-                    borderSide: BorderSide(color: colorScheme.primary, width: 2),
-                  ),
-                  contentPadding: const EdgeInsets.symmetric(
-                    horizontal: 12,
-                    vertical: 10,
-                  ),
-                  isDense: true,
-                ),
-                onSubmitted: (_) => _handleSubmit(),
-              ),
-              const SizedBox(height: 10),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  if (widget.isEditing && widget.onDelete != null)
-                    TextButton(
-                      onPressed: widget.onDelete,
-                      style: TextButton.styleFrom(
-                        foregroundColor: colorScheme.error,
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 12,
-                          vertical: 8,
-                        ),
-                      ),
-                      child: const Text(
-                        'Delete',
-                        style: TextStyle(fontSize: 13),
-                      ),
-                    ),
-                  const Spacer(),
-                  TextButton(
-                    onPressed: widget.onCancel,
-                    style: TextButton.styleFrom(
-                      foregroundColor: colorScheme.onSurfaceVariant,
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 12,
-                        vertical: 8,
-                      ),
-                    ),
-                    child: const Text(
-                      'Cancel',
-                      style: TextStyle(fontSize: 13),
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                  FilledButton(
-                    onPressed: _handleSubmit,
-                    style: FilledButton.styleFrom(
-                      backgroundColor: colorScheme.primary,
-                      foregroundColor: colorScheme.onPrimary,
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 16,
-                        vertical: 8,
-                      ),
-                    ),
-                    child: const Text(
-                      'Submit',
-                      style: TextStyle(fontSize: 13),
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
+// Re-export from bond_chat_ui package.
+export 'package:bond_chat_ui/bond_chat_ui.dart' show FeedbackDialog;
