+++ flutterui/lib/presentation/screens/chat/widgets/file_card.dart
@@ -1,233 +1,2 @@
-import 'dart:convert';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutterui/providers/services/service_providers.dart';
-
-class FileCard extends ConsumerStatefulWidget {
-  final String fileDataJson;
-
-  const FileCard({
-    super.key,
-    required this.fileDataJson,
-  });
-
-  @override
-  ConsumerState<FileCard> createState() => _FileCardState();
-}
-
-class _FileCardState extends ConsumerState<FileCard> {
-  bool _isDownloading = false;
-
-  Map<String, dynamic> _parseFileData() {
-    try {
-      return json.decode(widget.fileDataJson) as Map<String, dynamic>;
-    } catch (e) {
-      return {};
-    }
-  }
-
-  IconData _getFileIcon(String? mimeType) {
-    if (mimeType == null) return Icons.insert_drive_file;
-
-    if (mimeType.startsWith('image/')) return Icons.image;
-    if (mimeType.startsWith('video/')) return Icons.video_file;
-    if (mimeType.startsWith('audio/')) return Icons.audio_file;
-    if (mimeType == 'application/pdf') return Icons.picture_as_pdf;
-    if (mimeType.contains('text/')) return Icons.description;
-    if (mimeType.contains('spreadsheet') || mimeType.contains('excel') || mimeType == 'text/csv') {
-      return Icons.table_chart;
-    }
-    if (mimeType.contains('document') || mimeType.contains('word')) {
-      return Icons.article;
-    }
-    if (mimeType.contains('zip') || mimeType.contains('archive')) {
-      return Icons.folder_zip;
-    }
-    if (mimeType.contains('json') || mimeType.contains('xml')) {
-      return Icons.code;
-    }
-
-    return Icons.insert_drive_file;
-  }
-
-  String _formatFileSize(int? bytes) {
-    if (bytes == null) return 'Unknown size';
-
-    if (bytes < 1024) return '$bytes B';
-    if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
-    if (bytes < 1024 * 1024 * 1024) {
-      return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
-    }
-    return '${(bytes / (1024 * 1024 * 1024)).toStringAsFixed(1)} GB';
-  }
-
-  String _getFileTypeLabel(String? mimeType) {
-    if (mimeType == null) return 'File';
-
-    if (mimeType == 'application/pdf') return 'PDF';
-    if (mimeType == 'text/csv') return 'CSV';
-    if (mimeType.contains('spreadsheet') || mimeType.contains('excel')) return 'Excel';
-    if (mimeType.contains('document') || mimeType.contains('word')) return 'Word';
-    if (mimeType.contains('zip')) return 'ZIP';
-    if (mimeType.startsWith('image/')) return 'Image';
-    if (mimeType.startsWith('video/')) return 'Video';
-    if (mimeType.startsWith('audio/')) return 'Audio';
-    if (mimeType.startsWith('text/')) return 'Text';
-
-    return 'File';
-  }
-
-  Future<void> _downloadFile(String fileId, String fileName) async {
-    setState(() {
-      _isDownloading = true;
-    });
-
-    try {
-      final fileService = ref.read(fileServiceProvider);
-      await fileService.downloadFile(fileId, fileName);
-
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('Downloaded: $fileName'),
-            backgroundColor: Colors.green,
-          ),
-        );
-      }
-    } catch (e) {
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('Download failed: ${e.toString()}'),
-            backgroundColor: Colors.red,
-          ),
-        );
-      }
-    } finally {
-      if (mounted) {
-        setState(() {
-          _isDownloading = false;
-        });
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final fileData = _parseFileData();
-    final fileName = fileData['file_name'] as String? ?? 'Unknown File';
-    final fileSize = fileData['file_size'] as int?;
-    final mimeType = fileData['mime_type'] as String?;
-    final fileId = fileData['file_id'] as String?;
-
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-
-    return Container(
-      constraints: const BoxConstraints(maxWidth: 350),
-      decoration: BoxDecoration(
-        color: colorScheme.surfaceContainerHighest.withValues(alpha: 0.3),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(
-          color: colorScheme.outlineVariant,
-          width: 1,
-        ),
-      ),
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          borderRadius: BorderRadius.circular(12),
-          onTap: fileId != null && !_isDownloading
-              ? () => _downloadFile(fileId, fileName)
-              : null,
-          child: Padding(
-            padding: const EdgeInsets.all(12.0),
-            child: Row(
-              children: [
-                // File icon
-                Container(
-                  width: 48,
-                  height: 48,
-                  decoration: BoxDecoration(
-                    color: colorScheme.primaryContainer,
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: Icon(
-                    _getFileIcon(mimeType),
-                    color: colorScheme.onPrimaryContainer,
-                    size: 28,
-                  ),
-                ),
-                const SizedBox(width: 12),
-                // File info
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      Text(
-                        fileName,
-                        style: TextStyle(
-                          fontSize: 14,
-                          fontWeight: FontWeight.w500,
-                          color: colorScheme.onSurface,
-                        ),
-                        maxLines: 2,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      const SizedBox(height: 4),
-                      Row(
-                        children: [
-                          Container(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 6,
-                              vertical: 2,
-                            ),
-                            decoration: BoxDecoration(
-                              color: colorScheme.secondaryContainer,
-                              borderRadius: BorderRadius.circular(4),
-                            ),
-                            child: Text(
-                              _getFileTypeLabel(mimeType),
-                              style: TextStyle(
-                                fontSize: 10,
-                                fontWeight: FontWeight.w500,
-                                color: colorScheme.onSecondaryContainer,
-                              ),
-                            ),
-                          ),
-                          const SizedBox(width: 6),
-                          Text(
-                            _formatFileSize(fileSize),
-                            style: TextStyle(
-                              fontSize: 12,
-                              color: colorScheme.onSurfaceVariant,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-                const SizedBox(width: 8),
-                // Download button
-                if (_isDownloading)
-                  const SizedBox(
-                    width: 24,
-                    height: 24,
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  )
-                else
-                  Icon(
-                    Icons.download,
-                    color: colorScheme.primary,
-                    size: 24,
-                  ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
+// Re-export from bond_chat_ui package.
+export 'package:bond_chat_ui/bond_chat_ui.dart' show FileCard;
